Crate cosmic_settings_daemon

DBus interface proxy for: com.system76.CosmicSettingsDaemon

This code was generated by zbus-xmlgen 3.1.1 from DBus introspection data. Source: Interface '/com/system76/CosmicSettingsDaemon' from service 'com.system76.CosmicSettingsDaemon' on session bus.

You may prefer to adapt it, instead of using it verbatim.

More information can be found in the Writing a client proxy section of the zbus documentation.

This DBus object implements standard DBus interfaces, (org.freedesktop.DBus.*) for which the following zbus proxies can be used:

zbus::fdo::PeerProxy

zbus::fdo::PropertiesProxy

zbus::fdo::IntrospectableProxy

…consequently zbus-xmlgen did not generate code for the above interfaces.

Structs

Changed

A Changed signal.

ChangedArgs

Changed signal arguments.

ChangedIterator

A Iterator implementation that yields Changed signals.

ChangedStream

A Stream implementation that yields Changed signals.

ConfigProxy

ConfigProxyBlocking

CosmicSettingsDaemonProxy

CosmicSettingsDaemonProxyBlocking