cosmic_settings_daemon

Struct CosmicSettingsDaemonProxyBlocking

pub struct CosmicSettingsDaemonProxyBlocking<'p>(/* private fields */);

Implementations

impl<'p> CosmicSettingsDaemonProxyBlocking<'p>

pub fn new(conn: &Connection) -> Result<CosmicSettingsDaemonProxyBlocking<'p>>

Creates a new proxy with the default service and path.

pub fn builder(conn: &Connection) -> Builder<'p, Self>

Returns a customizable builder for this proxy.

pub fn into_inner(self) -> Proxy<'p>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'p>

The reference to the underlying zbus::Proxy.

pub fn inner_mut(&mut self) -> &mut Proxy<'p>

The mutable reference to the underlying zbus::Proxy.

pub fn decrease_display_brightness(&self) -> Result<()>

DecreaseDisplayBrightness method

pub fn decrease_keyboard_brightness(&self) -> Result<()>

DecreaseKeyboardBrightness method

pub fn increase_display_brightness(&self) -> Result<()>

IncreaseDisplayBrightness method

pub fn increase_keyboard_brightness(&self) -> Result<()>

IncreaseKeyboardBrightness method

pub fn watch_config( &self, id: &str, version: u64, ) -> Result<(OwnedObjectPath, OwnedWellKnownName)>

WatchConfig method

pub fn watch_state( &self, id: &str, version: u64, ) -> Result<(OwnedObjectPath, OwnedWellKnownName)>

WatchState method

pub fn display_brightness(&self) -> Result<i32>

DisplayBrightness property

pub fn cached_display_brightness( &self, ) -> Result<Option<<Result<i32> as ResultAdapter>::Ok>, <Result<i32> as ResultAdapter>::Err>

DisplayBrightness property Get the cached value of the DisplayBrightness property, or None if the property is not cached.

pub fn receive_display_brightness_changed( &self, ) -> PropertyIterator<'p, <Result<i32> as ResultAdapter>::Ok>

DisplayBrightness property Create a stream for the DisplayBrightness property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn set_display_brightness(&self, value: i32) -> Result<()>

pub fn keyboard_brightness(&self) -> Result<i32>

KeyboardBrightness property

pub fn cached_keyboard_brightness( &self, ) -> Result<Option<<Result<i32> as ResultAdapter>::Ok>, <Result<i32> as ResultAdapter>::Err>

KeyboardBrightness property Get the cached value of the KeyboardBrightness property, or None if the property is not cached.

pub fn receive_keyboard_brightness_changed( &self, ) -> PropertyIterator<'p, <Result<i32> as ResultAdapter>::Ok>

KeyboardBrightness property Create a stream for the KeyboardBrightness property changes. This is a convenient wrapper around zbus::blocking::Proxy::receive_property_changed.

pub fn set_keyboard_brightness(&self, value: i32) -> Result<()>